requirements.txt
matplotlib
numpy
sympy
mpld3
plotly
pillow